requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
anyio==4.3.0; python_version >= '3.8'
async-timeout==4.0.3; python_full_version < '3.12.0'
asyncpg==0.29.0; python_full_version >= '3.8.0'
attrs==23.2.0; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
frozenlist==1.4.1; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
multidict==6.0.5; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
python-telegram-bot==21.1.1; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.29; python_version >= '3.7'
typing-extensions==4.11.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
